+++ data-migrations/2026-08-04-1205-alter-questions.sql
@@ -1,0 +1,4 @@
+-- Added to support display logic for questions
+ALTER TABLE questions
+  ADD COLUMN displayLogicAction varchar(255) NOT NULL DEFAULT 'SHOW_QUESTION' AFTER required,
+  ADD COLUMN displayLogicMatchType varchar(255) NOT NULL DEFAULT 'ANY' AFTER displayLogicAction;